Moved files in components for organization. Also changed img to Image in ProductList.tsx and StoreStory.tsx (both are within components folder)
