+++ app/mystore/[slug]/address/page.js
@@ -1,14 +1,14 @@
-import React from 'react'
-import EditStoreAddressForm from '@/components/EditStoreAddressForm'
-import { getMyStoreInfo } from '@/utils/actions/storeActions'
+import React from "react";
+import EditStoreAddressForm from "@/components/edit-forms/EditStoreAddressForm";
+import { getMyStoreInfo } from "@/utils/actions/storeActions";
 
 const editStoreAddressPage = async ({ params }) => {
-	const storeInfo = await getMyStoreInfo(params.slug)
-	return (
-		<>
-			<EditStoreAddressForm storeInfo={storeInfo} myStore={params.slug} />
-		</>
-	)
-}
+  const storeInfo = await getMyStoreInfo(params.slug);
+  return (
+    <>
+      <EditStoreAddressForm storeInfo={storeInfo} myStore={params.slug} />
+    </>
+  );
+};
 
-export default editStoreAddressPage
+export default editStoreAddressPage;
